Renames session.js and changeController -> changeControllerTo
