+++ src/main/webapp/scripts/session.js
